backend/shared/requirements.txt
fastapi==0.110.0
pydantic==2.6.0
pydantic-settings==2.2.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.27.1
httpx==0.27.0
boto3==1.34.49
python-multipart==0.0.9